Update project configurations and version number
The commit modifies three files: Smpp.Net.Lib.csproj, common.props, and Smpp.Net.Client.csproj. The main changes were to adjust the TargetFramework to netstandard2.0 exclusively instead of net40, net48, and netstandard2.0. This was done to streamline the framework options and simplify deployment.

A version number in common.props was also changed from 1.10.0 to 2023.10.0, likely corresponding with a large-scale update or release.

Other minor updates include reordering of authors and owners, adjustments to url, and modification in the condition for the 'Release' configuration in Smpp.Net.Client.csproj file.