Update project details and dependencies in Elyfe.Smpp
Removed hardcoded version and updated information about the company, summary, and the inclusion of MinVer package. The changes were made to reflect the transfer of the project to a different company and to automate versioning. Adding the MinVer package allowed us to dynamically generate the version based on the Git history, removes the need to manually set it. In addition, updated project summary to more accurately represent the project's scope and purpose.