📦🔼 Update package versions and add MinVer
This commit updated the versions of packages used in the `Smpp.Net.Lib.csproj`, `Winton.Extensions.Configuration.Consul.csproj`, and `Smpp.Net.Client.csproj` projects. It also added a new package reference to MinVer. This helps us to ensure we're using the latest and most secure versions of our dependencies, and to specify the minimum version of a package that is required.